tui: rearrange buffer parts
