+++ mudpuppy/src/tui/layout.rs
@@ -1,19 +1,11 @@
 use std::collections::HashMap;
-use std::fmt;
-use std::fmt::{Display, Formatter};
 
 use pyo3::types::{PyAnyMethods, PyDict, PyList, PyListMethods, PyTuple};
-use pyo3::{pyclass, pymethods, Bound, Py, PyRef, Python};
+use pyo3::{pyclass, pymethods, Bound, Py, Python};
 use ratatui::layout::{Constraint, Direction, Layout, Rect};
-use ratatui::widgets::{Borders, Clear};
-use ratatui::Frame;
-use tracing::trace;
 
-use crate::client::output::Output;
 use crate::error::Error;
-use crate::tui::buffer;
-use crate::tui::buffer::DrawScrollbar;
-use crate::{idmap, Result};
+use crate::Result;
 
 // TODO(XXX): Rename?
 #[derive(Debug, Clone)]
@@ -348,284 +340,5 @@ impl PyConstraint {
 
     fn __repr__(&self) -> String {
         self.__str__()
-    }
-}
-
-/// An "Extra Buffer" is a user created buffer that can be used to display
-/// arbitrary content.
-///
-/// It's similar to `MudBuffer` and `ScrollWindow`, but without being tailored
-/// to one of those two specific tasks.
-#[derive(Debug, Clone)]
-#[pyclass]
-pub struct ExtraBuffer {
-    #[pyo3(get)]
-    pub id: u32,
-
-    #[pyo3(get)]
-    pub config: Py<BufferConfig>,
-}
-
-impl ExtraBuffer {
-    /// # Errors
-    /// TODO(XXX): docs.
-    pub fn draw_buffer(
-        &mut self,
-        f: &mut Frame<'_>,
-        sections: &HashMap<String, Rect>,
-    ) -> Result<()> {
-        Python::with_gil(|py| {
-            let mut config: BufferConfig = self.config.extract(py)?;
-            let buffer_area = sections
-                .get(&config.layout_name)
-                .ok_or(Error::LayoutMissing(config.layout_name.clone()))?;
-
-            let mut output: Output = config.output.extract(py)?;
-
-            // Make sure to clear the viewport first - we might be drawing on top of already
-            // rendered content.
-            f.render_widget(Clear, *buffer_area);
-
-            buffer::draw(
-                f,
-                &mut config,
-                output.read_received().iter(),
-                |_| true,
-                buffer_area,
-                // TODO(XXX): config for scroll bar render...
-                DrawScrollbar::Always,
-            )
-        })
-    }
-}
-
-#[pymethods]
-impl ExtraBuffer {
-    fn __str__(&self, py: Python<'_>) -> String {
-        let config: PyRef<'_, BufferConfig> = self.config.extract(py).unwrap();
-        format!("Buffer({}) config: {}", self.id, *config)
-    }
-
-    fn __repr__(&self) -> String {
-        format!("{self:?}")
-    }
-}
-
-impl Display for ExtraBuffer {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "Buffer({})", self.id)
-    }
-}
-
-impl idmap::Identifiable for ExtraBuffer {
-    fn id(&self) -> u32 {
-        self.id
-    }
-}
-
-#[derive(Debug, Clone)]
-#[pyclass]
-#[allow(clippy::struct_excessive_bools)] // TODO(XXX): Consider.
-pub struct BufferConfig {
-    #[pyo3(get, set)]
-    pub layout_name: String,
-
-    #[pyo3(get, set)]
-    pub line_wrap: bool,
-
-    #[pyo3(get, set)]
-    pub border_top: bool,
-
-    #[pyo3(get, set)]
-    pub border_bottom: bool,
-
-    #[pyo3(get, set)]
-    pub border_left: bool,
-
-    #[pyo3(get, set)]
-    pub border_right: bool,
-
-    #[pyo3(get, set)]
-    pub direction: BufferDirection,
-
-    #[pyo3(get)]
-    pub output: Py<Output>,
-
-    #[pyo3(get)]
-    pub scroll_pos: usize,
-
-    #[pyo3(get)]
-    pub max_scroll: usize,
-}
-
-impl BufferConfig {
-    #[must_use]
-    pub fn area_inside_borders(&self, mut area: Rect, scrollbar: bool) -> Rect {
-        if self.border_top {
-            area.height = area.height.saturating_sub(1);
-            area.y = area.y.saturating_add(1);
-        }
-        if self.border_bottom {
-            area.height = area.height.saturating_sub(1);
-        }
-        if self.border_left {
-            area.width = area.width.saturating_sub(1);
-            area.x = area.x.saturating_add(1);
-        }
-        if self.border_right {
-            area.width = area.width.saturating_sub(1);
-        }
-        if scrollbar {
-            area.width = area.width.saturating_sub(1);
-        }
-        area
-    }
-
-    #[must_use]
-    pub fn area_inside_top_borders(&self, mut area: Rect) -> Rect {
-        if self.border_top {
-            area.height = area.height.saturating_sub(1);
-            area.y = area.y.saturating_add(1);
-        }
-        if self.border_bottom {
-            area.height = area.height.saturating_sub(1);
-        }
-        area
-    }
-
-    #[must_use]
-    pub fn borders(&self) -> Borders {
-        let mut borders = Borders::empty();
-        if self.border_top {
-            borders |= Borders::TOP;
-        }
-        if self.border_bottom {
-            borders |= Borders::BOTTOM;
-        }
-        if self.border_left {
-            borders |= Borders::LEFT;
-        }
-        if self.border_right {
-            borders |= Borders::RIGHT;
-        }
-        borders
-    }
-}
-
-#[pymethods]
-impl BufferConfig {
-    /// # Errors
-    /// If the layout name is empty
-    #[new]
-    pub fn new(layout_name: String) -> Result<Self> {
-        if layout_name.is_empty() {
-            return Err(Error::BadLayout);
-        }
-        let output = Python::with_gil(|py| Py::new(py, Output::new()))?;
-        Ok(Self {
-            layout_name,
-            line_wrap: false,
-            output,
-            border_top: false,
-            border_bottom: false,
-            border_left: false,
-            border_right: false,
-            direction: BufferDirection::default(),
-            scroll_pos: 0,
-            max_scroll: 0,
-        })
-    }
-
-    #[must_use]
-    pub fn scroll(&self) -> usize {
-        self.scroll_pos
-    }
-
-    pub fn scroll_up(&mut self, lines: u16) {
-        trace!("scrolling up: scroll-pos: {}", self.scroll_pos);
-        self.scroll_pos = self
-            .scroll_pos
-            .checked_add(lines as usize)
-            .unwrap_or(self.scroll_pos);
-        trace!("scrolling up: scroll-pos now {}", self.scroll_pos);
-    }
-
-    pub fn scroll_down(&mut self, lines: u16) {
-        trace!("scrolling down: scroll-pos: {}", self.scroll_pos);
-        self.scroll_pos = self.scroll_pos.saturating_sub(lines as usize);
-        trace!("scrolling down: scroll-pos now {}", self.scroll_pos);
-    }
-
-    pub fn scroll_bottom(&mut self) {
-        trace!("scrolling to bottom: scroll-pos: {}", self.scroll_pos);
-        self.scroll_pos = 1;
-        trace!("scrolling to bottom: scroll-pos now {}", self.scroll_pos);
-    }
-
-    pub fn scroll_to(&mut self, scroll: usize) {
-        trace!(
-            "scrolling to pos: scroll-pos {}: {}",
-            scroll,
-            self.scroll_pos
-        );
-        self.scroll_pos = scroll;
-        trace!(
-            "scrolling to pos: scroll-pos {} now: {}",
-            scroll,
-            self.scroll_pos
-        );
-    }
-
-    pub fn scroll_max(&mut self) {
-        trace!("scrolling to max: scroll-pos: {}", self.max_scroll);
-        self.scroll_pos = self.max_scroll;
-        trace!("scrolling to max: scroll-pos now: {}", self.scroll_pos);
-    }
-
-    #[allow(clippy::trivially_copy_pass_by_ref)]
-    fn __str__(&self) -> String {
-        // TODO(XXX): nicer str format
-        format!("{self:?}")
-    }
-
-    #[allow(clippy::trivially_copy_pass_by_ref)]
-    fn __repr__(&self) -> String {
-        format!("{self:?}")
-    }
-}
-
-impl Display for BufferConfig {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "BufferConfig({})", self.layout_name)
-    }
-}
-
-#[derive(Debug, Default, Clone, Copy, Eq, PartialEq)]
-#[pyclass(eq, eq_int)]
-pub enum BufferDirection {
-    TopToBottom,
-    #[default]
-    BottomToTop,
-}
-
-#[pymethods]
-impl BufferDirection {
-    #[allow(clippy::trivially_copy_pass_by_ref)]
-    fn __str__(&self) -> String {
-        format!("{self}")
-    }
-
-    #[allow(clippy::trivially_copy_pass_by_ref)]
-    fn __repr__(&self) -> String {
-        format!("{self:?}")
-    }
-}
-
-impl Display for BufferDirection {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match self {
-            BufferDirection::TopToBottom => write!(f, "top to bottom"),
-            BufferDirection::BottomToTop => write!(f, "bottom to top"),
-        }
     }
 }
