+++ mudpuppy/src/tui/mod.rs
@@ -1,4 +1,5 @@
 pub mod buffer;
+pub(crate) mod extrabuffer;
 mod input;
 pub mod layout;
 pub mod mudbuffer;
